refactor(sdk-dotnet): move public API to root namespace Envilder
BREAKING CHANGE: Public types moved from Envilder.Application and
Envilder.Domain to namespace Envilder. Extension methods moved to
Microsoft.Extensions.Configuration and
Microsoft.Extensions.DependencyInjection respectively.

Consumers now only need 'using Envilder;' (or 'using static Envilder.Envilder;')
instead of multiple namespace imports.
