+++ src/sdks/dotnet/Application/MapFileParser.cs
@@ -1,6 +1,5 @@
-namespace Envilder.Application;
+namespace Envilder;
 
-using global::Envilder.Domain;
 using System.Collections.Generic;
 using System.Text.Json;
 using System.Text.Json.Serialization;
